requirements.txt
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
